#24 Separate errorHandler to apiErrorHandler and uncaughtErrorHandler
